ZC-10593: Merge last upstream bit into ea-nginx

+++ SOURCES/ngx_http_pipelog_module-ngx_http_pipelog_module.c
@@ -12,7 +12,6 @@
 #if (NGX_ZLIB)
 #include <zlib.h>
 #endif
-
 #include <wordexp.h>
 
 typedef struct ngx_http_log_op_s  ngx_http_log_op_t;
@@ -89,6 +88,11 @@ typedef struct {
 } ngx_http_log_var_t;
 
 
+#define NGX_HTTP_LOG_ESCAPE_DEFAULT  0
+#define NGX_HTTP_LOG_ESCAPE_JSON     1
+#define NGX_HTTP_LOG_ESCAPE_NONE     2
+
+
 static void ngx_http_log_write(ngx_http_request_t *r, ngx_http_log_t *log,
     u_char *buf, size_t len);
 static ssize_t ngx_http_log_script_write(ngx_http_request_t *r,
@@ -125,12 +129,20 @@ static u_char *ngx_http_log_request_length(ngx_http_request_t *r, u_char *buf,
     ngx_http_log_op_t *op);
 
 static ngx_int_t ngx_http_log_variable_compile(ngx_conf_t *cf,
-    ngx_http_log_op_t *op, ngx_str_t *value);
+    ngx_http_log_op_t *op, ngx_str_t *value, ngx_uint_t escape);
 static size_t ngx_http_log_variable_getlen(ngx_http_request_t *r,
     uintptr_t data);
 static u_char *ngx_http_log_variable(ngx_http_request_t *r, u_char *buf,
     ngx_http_log_op_t *op);
 static uintptr_t ngx_http_log_escape(u_char *dst, u_char *src, size_t size);
+static size_t ngx_http_log_json_variable_getlen(ngx_http_request_t *r,
+    uintptr_t data);
+static u_char *ngx_http_log_json_variable(ngx_http_request_t *r, u_char *buf,
+    ngx_http_log_op_t *op);
+static size_t ngx_http_log_unescaped_variable_getlen(ngx_http_request_t *r,
+    uintptr_t data);
+static u_char *ngx_http_log_unescaped_variable(ngx_http_request_t *r,
+    u_char *buf, ngx_http_log_op_t *op);
 
 
 static void *ngx_http_log_create_main_conf(ngx_conf_t *cf);
@@ -146,7 +158,7 @@ static char *ngx_http_log_compile_format(ngx_conf_t *cf,
 static char *ngx_http_log_open_file_cache(ngx_conf_t *cf, ngx_command_t *cmd,
     void *conf);
 static ngx_int_t ngx_http_log_init(ngx_conf_t *cf);
-
+static void close_pipes(ngx_cycle_t *cycle);
 
 static ngx_command_t  ngx_http_log_commands[] = {
 
@@ -850,7 +862,7 @@ ngx_http_log_request_length(ngx_http_request_t *r, u_char *buf,
 
 static ngx_int_t
 ngx_http_log_variable_compile(ngx_conf_t *cf, ngx_http_log_op_t *op,
-    ngx_str_t *value)
+    ngx_str_t *value, ngx_uint_t escape)
 {
     ngx_int_t  index;
 
@@ -860,8 +872,23 @@ ngx_http_log_variable_compile(ngx_conf_t *cf, ngx_http_log_op_t *op,
     }
 
     op->len = 0;
-    op->getlen = ngx_http_log_variable_getlen;
-    op->run = ngx_http_log_variable;
+
+    switch (escape) {
+    case NGX_HTTP_LOG_ESCAPE_JSON:
+        op->getlen = ngx_http_log_json_variable_getlen;
+        op->run = ngx_http_log_json_variable;
+        break;
+
+    case NGX_HTTP_LOG_ESCAPE_NONE:
+        op->getlen = ngx_http_log_unescaped_variable_getlen;
+        op->run = ngx_http_log_unescaped_variable;
+        break;
+
+    default: /* NGX_HTTP_LOG_ESCAPE_DEFAULT */
+        op->getlen = ngx_http_log_variable_getlen;
+        op->run = ngx_http_log_variable;
+    }
+
     op->data = index;
 
     return NGX_OK;
@@ -966,6 +993,80 @@ ngx_http_log_escape(u_char *dst, u_char *src, size_t size)
     }
 
     return (uintptr_t) dst;
+}
+
+
+static size_t
+ngx_http_log_json_variable_getlen(ngx_http_request_t *r, uintptr_t data)
+{
+    uintptr_t                   len;
+    ngx_http_variable_value_t  *value;
+
+    value = ngx_http_get_indexed_variable(r, data);
+
+    if (value == NULL || value->not_found) {
+        return 0;
+    }
+
+    len = ngx_escape_json(NULL, value->data, value->len);
+
+    value->escape = len ? 1 : 0;
+
+    return value->len + len;
+}
+
+
+static u_char *
+ngx_http_log_json_variable(ngx_http_request_t *r, u_char *buf,
+    ngx_http_log_op_t *op)
+{
+    ngx_http_variable_value_t  *value;
+
+    value = ngx_http_get_indexed_variable(r, op->data);
+
+    if (value == NULL || value->not_found) {
+        return buf;
+    }
+
+    if (value->escape == 0) {
+        return ngx_cpymem(buf, value->data, value->len);
+
+    } else {
+        return (u_char *) ngx_escape_json(buf, value->data, value->len);
+    }
+}
+
+
+static size_t
+ngx_http_log_unescaped_variable_getlen(ngx_http_request_t *r, uintptr_t data)
+{
+    ngx_http_variable_value_t  *value;
+
+    value = ngx_http_get_indexed_variable(r, data);
+
+    if (value == NULL || value->not_found) {
+        return 0;
+    }
+
+    value->escape = 0;
+
+    return value->len;
+}
+
+
+static u_char *
+ngx_http_log_unescaped_variable(ngx_http_request_t *r, u_char *buf,
+    ngx_http_log_op_t *op)
+{
+    ngx_http_variable_value_t  *value;
+
+    value = ngx_http_get_indexed_variable(r, op->data);
+
+    if (value == NULL || value->not_found) {
+        return buf;
+    }
+
+    return ngx_cpymem(buf, value->data, value->len);
 }
 
 
@@ -1389,11 +1490,30 @@ ngx_http_log_compile_format(ngx_conf_t *cf, ngx_array_t *flushes,
     size_t               i, len;
     ngx_str_t           *value, var;
     ngx_int_t           *flush;
-    ngx_uint_t           bracket;
+    ngx_uint_t           bracket, escape;
     ngx_http_log_op_t   *op;
     ngx_http_log_var_t  *v;
 
+    escape = NGX_HTTP_LOG_ESCAPE_DEFAULT;
     value = args->elts;
+
+    if (s < args->nelts && ngx_strncmp(value[s].data, "escape=", 7) == 0) {
+        data = value[s].data + 7;
+
+        if (ngx_strcmp(data, "json") == 0) {
+            escape = NGX_HTTP_LOG_ESCAPE_JSON;
+
+        } else if (ngx_strcmp(data, "none") == 0) {
+            escape = NGX_HTTP_LOG_ESCAPE_NONE;
+
+        } else if (ngx_strcmp(data, "default") != 0) {
+            ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
+                               "unknown log format escaping \"%s\"", data);
+            return NGX_CONF_ERROR;
+        }
+
+        s++;
+    }
 
     for ( /* void */ ; s < args->nelts; s++) {
 
@@ -1473,7 +1593,9 @@ ngx_http_log_compile_format(ngx_conf_t *cf, ngx_array_t *flushes,
                     }
                 }
 
-                if (ngx_http_log_variable_compile(cf, op, &var) != NGX_OK) {
+                if (ngx_http_log_variable_compile(cf, op, &var, escape)
+                    != NGX_OK)
+                {
                     return NGX_CONF_ERROR;
                 }
 
@@ -1717,8 +1839,8 @@ typedef struct {
 
 typedef struct {
     ngx_fd_t fd[2];
-    ngx_str_t command;
     ngx_uint_t nonblocking;
+    wordexp_t  w;
     pid_t pid;
     struct timeval timestamp;
 } ngx_http_pipelog_pim_t;
@@ -1948,21 +2070,19 @@ ngx_http_pipelog_set_log (ngx_conf_t *cf, ngx_command_t *cmd, void *conf) {
     if (pipelog->pim->nonblocking) {
         ngx_nonblocking(pipelog->pim->fd[1]);
     }
-    pipelog->pim->command = value[1];
 
-    wordexp_t  w;
-    int result = wordexp((const char*)value[1].data, &w, WRDE_NOCMD);
+    int result = wordexp((const char*)value[1].data, &pipelog->pim->w, 0);
     if (result != 0) {
         ngx_conf_log_error(NGX_LOG_EMERG, cf, 0, "wrong command syntax(%i): %s", result, value[1].data);
         return NGX_CONF_ERROR;
     }
-    int word_count = w.we_wordc;
-    wordfree(&w);
+    int word_count = pipelog->pim->w.we_wordc;
 
     if (word_count < 1) {
         ngx_conf_log_error(NGX_LOG_EMERG, cf, 0, "no command: %s", value[1].data);
         return NGX_CONF_ERROR;
     }
+
     return NGX_CONF_OK;
 }
 
@@ -2120,45 +2240,29 @@ ngx_http_pipelog_init (ngx_conf_t *cf) {
 }
 
 static ngx_pid_t
-ngx_http_pipelog_command_exec (ngx_str_t *command, ngx_fd_t rfd, ngx_cycle_t *cycle, sigset_t mask) {
+ngx_http_pipelog_command_exec (wordexp_t *w, ngx_fd_t rfd, ngx_cycle_t *cycle) {
     ngx_pid_t pid;
-    char cmd[1024];
     ngx_fd_t fd;
-    wordexp_t w;
-    if (command->len > sizeof(cmd) - 1) {
+    sigset_t mask;
+
+    if (!w) {
         return -1;
     }
-
     pid = fork();
     switch (pid){
     case -1:
         ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: ngx_http_pipelog_command_exec fork(): error", MODULE_NAME);
         return NGX_ERROR;
     case 0:
-        if(sigprocmask(SIG_SETMASK, &mask, NULL) == -1) {
-            ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: ngx_http_pipelog_command_exec sigprocmask(): error", MODULE_NAME);
-        }
         dup2(rfd, STDIN_FILENO);
         for (fd = 0; fd < NOFILE; fd++) {
             if (fd != STDIN_FILENO && fd != STDOUT_FILENO && fd != STDERR_FILENO) {
                 close(fd);
             }
         }
-        memset(cmd, 0, sizeof(cmd));
-        memcpy(cmd, command->data, command->len);
-        if (wordexp(cmd, &w, WRDE_NOCMD) == 0) {
-            if (w.we_wordc > 0) {
-                char *bin = w.we_wordv[0];
-                if (*bin) {
-                    /* We need to unblock signals for execve() so that
-                       processes can see SIGTERM/SIGHUP */
-                    sigfillset(&mask);
-                    sigprocmask(SIG_UNBLOCK, &mask, NULL);
-                    execve(bin, w.we_wordv, NULL);
-                }
-            }
-            wordfree(&w);
-        }
+        sigfillset(&mask);
+        sigprocmask(SIG_UNBLOCK, &mask, NULL);
+        execvp(w->we_wordv[0], w->we_wordv);
         exit(1);
     default:
         break;
@@ -2167,7 +2271,7 @@ ngx_http_pipelog_command_exec (ngx_str_t *command, ngx_fd_t rfd, ngx_cycle_t *cy
 }
 
 static ngx_uint_t
-ngx_http_pipelog_reap_child (ngx_cycle_t *cycle, sigset_t mask) {
+ngx_http_pipelog_reap_child (ngx_cycle_t *cycle) {
     struct timeval now, diff;
     ngx_http_pipelog_main_conf_t *pmcf;
     ngx_uint_t num, idx;
@@ -2197,7 +2301,7 @@ ngx_http_pipelog_reap_child (ngx_cycle_t *cycle, sigset_t mask) {
             continue;
         }
         pim[idx].timestamp = now;
-        pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].command, pim[idx].fd[0], cycle, mask);
+        pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].w, pim[idx].fd[0], cycle);
         if (pim[idx].pid == -1) {
             ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: reap child process (pid='%d'), respawn child process failed", MODULE_NAME, pid);
             continue;
@@ -2210,7 +2314,7 @@ ngx_http_pipelog_reap_child (ngx_cycle_t *cycle, sigset_t mask) {
 void
 ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
     struct sigaction sa;
-    sigset_t set, old;
+    sigset_t set;
     ngx_core_conf_t *ccf;
     struct timeval now, diff;
     ngx_http_pipelog_main_conf_t *pmcf;
@@ -2219,7 +2323,7 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
     struct timespec timeout;
     int sig;
 
-    ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0, "%s: start logger process %d", MODULE_NAME, getpid());
+    ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0, "p[%d]: %s: new logger process started", getpid(), MODULE_NAME);
 
     memset(&sa, 0, sizeof(sa));
 
@@ -2228,7 +2332,7 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
     sigemptyset(&set);
     sigaddset(&set, SIGTERM);
     sigaddset(&set, SIGCHLD);
-    sigprocmask(SIG_BLOCK, &set, &old);
+    sigprocmask(SIG_BLOCK, &set, NULL);
 
     ccf = (ngx_core_conf_t *)ngx_get_conf(cycle->conf_ctx, ngx_core_module);
     if (geteuid() == 0) {
@@ -2242,7 +2346,7 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
     pmcf = ngx_http_cycle_get_module_main_conf(cycle, ngx_http_pipelog_module);
     pim = pmcf->pims.elts;
     for (idx = 0; idx < pmcf->pims.nelts; idx++) {
-        pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].command, pim[idx].fd[0], cycle, old);
+        pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].w, pim[idx].fd[0], cycle);
         if (pim[idx].pid == -1) {
             ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: ngx_http_pipelog_command_exec(): error", MODULE_NAME);
         }
@@ -2256,6 +2360,7 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
 
         if(sig == SIGTERM) {
             ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: ngx_http_pipelog_command_exec(): SIGTERM detected, gracefully shutting down...", MODULE_NAME);
+            close_pipes(cycle);
             if(killpg(0, SIGTERM) == -1) {
                 ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: ngx_http_pipelog_command_exec: killpg(%d, SIGTERM) failed ", MODULE_NAME, 0);
             }
@@ -2263,7 +2368,7 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
         }
 
         if (sig != -1) {
-            ngx_http_pipelog_reap_child(cycle, old);
+            ngx_http_pipelog_reap_child(cycle);
         } else {
             gettimeofday(&now, NULL);
             pim = pmcf->pims.elts;
@@ -2276,13 +2381,29 @@ ngx_http_pipelog_logger_process_main (ngx_cycle_t *cycle) {
                     continue;
                 }
                 pim[idx].timestamp = now;
-                pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].command, pim[idx].fd[0], cycle, old);
+                pim[idx].pid = ngx_http_pipelog_command_exec(&pim[idx].w, pim[idx].fd[0], cycle);
                 if (pim[idx].pid == -1) {
                     ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: respawn child process failed", MODULE_NAME);
                     continue;
                 }
                 ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: respawn child process (pid='%d')", MODULE_NAME, pim[idx].pid);
             }
+        }
+    }
+}
+
+static void close_pipes(ngx_cycle_t *cycle) {
+    ngx_http_pipelog_main_conf_t *pmcf;
+    ngx_http_pipelog_pim_t *pim;
+    ngx_uint_t idx;
+
+    if (cycle && cycle->modules) {
+        pmcf = ngx_http_cycle_get_module_main_conf(cycle, ngx_http_pipelog_module);
+        pim = pmcf->pims.elts;
+        for (idx = 0; idx < pmcf->pims.nelts; idx++) {
+            ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0, "p[%d]: %s: closing pipes: rd: %i, wr: %i", getpid(), MODULE_NAME, pim[idx].fd[0], pim[idx].fd[1]);
+            close(pim[idx].fd[0]);
+            close(pim[idx].fd[1]);
         }
     }
 }
@@ -2302,7 +2423,7 @@ init_module (ngx_cycle_t *cycle) {
         case 0:
             if(setpgid(0, 0) == -1) {
                 ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: init_module: setpgid() failed", MODULE_NAME);
-                return NGX_ERROR;
+                exit(1);
             }
             fd = open("/dev/null", O_RDWR);
             if (fd != -1) {
@@ -2311,10 +2432,12 @@ init_module (ngx_cycle_t *cycle) {
                 dup2(fd, STDERR_FILENO);
                 close(fd);
             }
+            close_pipes(cycle->old_cycle);
             ngx_setproctitle(LOGGER_PROC_NAME);
             ngx_http_pipelog_logger_process_main(cycle);
             exit(1);
         default:
+            close_pipes(cycle->old_cycle);
             break;
         }
     }
@@ -2324,9 +2447,10 @@ init_module (ngx_cycle_t *cycle) {
 static void
 exit_process (ngx_cycle_t *cycle) {
     ngx_http_pipelog_main_conf_t *pmcf;
-
     pmcf = ngx_http_cycle_get_module_main_conf(cycle, ngx_http_pipelog_module);
     ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0, "%s: exit_process called", MODULE_NAME);
+    close_pipes(cycle);
+
     if(killpg(pmcf->pid, SIGTERM) == -1) {
         ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: exit_process: killpg(%d, SIGTERM) failed ", MODULE_NAME, pmcf->pid);
     }
@@ -2335,10 +2459,13 @@ exit_process (ngx_cycle_t *cycle) {
 static void
 exit_master (ngx_cycle_t *cycle) {
     ngx_http_pipelog_main_conf_t *pmcf;
-
     pmcf = ngx_http_cycle_get_module_main_conf(cycle, ngx_http_pipelog_module);
     ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0, "%s: exit_master called", MODULE_NAME);
-    if(kill(pmcf->pid, SIGKILL) == -1) {
-        ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: exit_master: kill(%d, SIGTERM) failed ", MODULE_NAME, pmcf->pid);
+    close_pipes(cycle);
+
+    if(!ngx_terminate) {
+        if(kill(pmcf->pid, SIGKILL) == -1) {
+            ngx_log_error(NGX_LOG_ALERT, cycle->log, 0, "%s: exit_master: kill(%d, SIGTERM) failed ", MODULE_NAME, pmcf->pid);
+        }
     }
 }
